Checking for errors in retrieving a session and creating a new one on err
This is different from the generateNewSession method, which is just an override for store.generate. The createSession method will call store.generate. This will also check for errors retrieving the session from the store and creating a new session if there isn't one in the store. This is primarily for edge cases where the client sends a cookie with a session Id that's valid but the session does not exist in the store